Sukurtas metodas readFiles
